Add files via upload
I created the new column  entitled, "poli_pref" by dividing the rows in the "votes_dem" column by the rows in the "votes_gop" column. I did not set up the calculation so that a "poli_pref" value of "1" represents a neutral county.  Instead, it was set up so that the higher the "poli_pref" value you is, the more Democratic the county is.